HDDS-10050 Move createKeys method and delete the empty StorageContainerManagerTestHelper
